Added POST Service for Country API

+++ src/main/java/com/reboot/admin/model/Country.java
@@ -4,14 +4,15 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class Country extends CommonModel {
-	private int id;
+	private long id;
 	private String name;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+
+	public void setId(long id) {
 		this.id = id;
 	}
